Add a number of org wide default templates
The `.github` project in an org is a special.  It allows you to set org
wide defaults for a number of template files.  There is no impact to
existing projects that have custom versions of these files but all
projects that don't have them will use the ones from this repo.

Related documentation:
* https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file

I took a few liberties that we may not want to add in right now.  In an
effort to encourage people to think about the changes they're
submitting, there are two checklists.

The first is to indicate the change type, which should be a single item.  (PRs with multiple changes
tend to be more complex, require longer review and reduce velocity.)

The second checklist is a way to encourage people to make sure they've
gone through all the required steps.  Lint, comments, testing, etc.

Feedback requested.